Split off 6FD410 based on bss
