+++ include/6FD190.h
@@ -14,10 +14,5 @@ void* func_800D96E8_jp(s32 arg0);
 s32 func_800D9700_jp(void);
 void* func_800D9750_jp(void);
 // void func_800D9760_jp();
-void func_800D9770_jp(void);
-void func_800D9788_jp(void);
-// void func_800D97A0_jp();
-// void func_800D97F8_jp();
-// void func_800D986C_jp();
 
 #endif
