+++ include/6FD410.h
@@ -1,0 +1,12 @@
+#ifndef C_6FD410_H
+#define C_6FD410_H
+
+#include "ultra64.h"
+
+void func_800D9770_jp(void);
+void func_800D9788_jp(void);
+// void func_800D97A0_jp();
+// void func_800D97F8_jp();
+// void func_800D986C_jp();
+
+#endif
