+++ linker_scripts/jp/symbol_addrs_code.txt
@@ -6209,10 +6209,10 @@ B_80146044_jp = 0x80146044; //
 B_80146050_jp = 0x80146050; //
 B_80146058_jp = 0x80146058; //
 B_80146060_jp = 0x80146060; // size:0xC
-B_8014606C_jp = 0x8014606C; //
-B_80146070_jp = 0x80146070; //
-B_80146080_jp = 0x80146080; //
-B_80146084_jp = 0x80146084; //
+B_8014606C_jp = 0x8014606C; // size:0x4
+B_80146070_jp = 0x80146070; // size:0x4
+B_80146080_jp = 0x80146080; // size:0x4
+B_80146084_jp = 0x80146084; // size:0x4
 B_FLT_80146090_jp = 0x80146090; // type:f32
 B_FLT_80146094_jp = 0x80146094; // type:f32
 B_FLT_80146098_jp = 0x80146098; // type:f32
