+++ yamls/jp/code.yaml
@@ -108,6 +108,7 @@
       - [0x6FB340, asm, code/6FB340]
       - [0x6FC540, asm, code/speed_meter]
       - [0x6FD190, c, code/6FD190]
+      - [0x6FD410, c, code/6FD410]
       - [0x6FD530, c, code/sys_initial_check]
       - [0x6FD840, asm, code/sys_math]
       - [0x6FDC40, asm, code/sys_math3d]
@@ -331,6 +332,9 @@
       - { start: 0x73F4D0, type: .bss, vram: 0x80145620, name: code/lb_rtc}
       - { start: 0x73F4D0, type: .bss, vram: 0x80145630, name: code/main}
       - { start: 0x73F4D0, type: bss, vram: 0x801458E0, name: code/801458E0}
+      - { start: 0x73F4D0, type: bss, vram: 0x80146060, name: code/6FD190}
+      - { start: 0x73F4D0, type: bss, vram: 0x80146080, name: code/6FD410}
+      - { start: 0x73F4D0, type: bss, vram: 0x80146090, name: code/80146090}
       - { start: 0x73F4D0, type: .bss, vram: 0x801462B0, name: code/sys_matrix }
       - { start: 0x73F4D0, type: .bss, vram: 0x801462C0, name: code/sys_stacks }
       - { start: 0x73F4D0, type: .bss, vram: 0x801492C0, name: code/zurumode }
